Move scrapeSlaveStatus to collector package.
