+++ mysqld_exporter.go
@@ -134,7 +134,6 @@ const (
 	exporter          = "exporter"
 	informationSchema = "info_schema"
 	performanceSchema = "perf_schema"
-	slaveStatus       = "slave_status"
 	binlog            = "binlog"
 	tokudb            = "engine_tokudb"
 )
@@ -142,7 +141,6 @@ const (
 // Metric SQL Queries.
 const (
 	sessionSettingsQuery       = `SET SESSION log_slow_filter = 'tmp_table_on_disk,filesort_on_disk'`
-	slaveStatusQuery           = `SHOW SLAVE STATUS`
 	binlogQuery                = `SHOW BINARY LOGS`
 	engineTokudbStatusQuery    = `SHOW ENGINE TOKUDB STATUS`
 	logbinQuery                = `SELECT @@log_bin`
@@ -289,8 +287,6 @@ const (
 		  FROM information_schema.query_response_time
 	`
 )
-
-var slaveStatusQuerySuffixes = [3]string{" NONBLOCKING", " NOLOCK", ""}
 
 // landingPage contains the HTML served at '/'.
 // TODO: Make this nicer and more informative.
@@ -815,7 +811,7 @@ func (e *Exporter) scrape(ch chan<- prometheus.Metric) {
 		}
 	}
 	if *collectSlaveStatus {
-		if err = scrapeSlaveStatus(db, ch); err != nil {
+		if err = collector.ScrapeSlaveStatus(db, ch); err != nil {
 			log.Errorln("Error scraping for collect.slave_status:", err)
 			e.scrapeErrors.WithLabelValues("collect.slave_status").Inc()
 		}
@@ -910,57 +906,6 @@ func (e *Exporter) scrape(ch chan<- prometheus.Metric) {
 			e.scrapeErrors.WithLabelValues("collect.engine_tokudb_status").Inc()
 		}
 	}
-}
-
-func scrapeSlaveStatus(db *sql.DB, ch chan<- prometheus.Metric) error {
-	var (
-		slaveStatusRows *sql.Rows
-		err             error
-	)
-	// Leverage lock-free SHOW SLAVE STATUS by guessing the right suffix
-	for _, suffix := range slaveStatusQuerySuffixes {
-		slaveStatusRows, err = db.Query(fmt.Sprint(slaveStatusQuery, suffix))
-		if err == nil {
-			break
-		}
-	}
-	if err != nil {
-		return err
-	}
-	defer slaveStatusRows.Close()
-
-	if slaveStatusRows.Next() {
-		// There is either no row in SHOW SLAVE STATUS (if this is not a
-		// slave server), or exactly one. In case of multi-source
-		// replication, things work very much differently. This code
-		// cannot deal with that case.
-		slaveCols, err := slaveStatusRows.Columns()
-		if err != nil {
-			return err
-		}
-
-		// As the number of columns varies with mysqld versions,
-		// and sql.Scan requires []interface{}, we need to create a
-		// slice of pointers to the elements of slaveData.
-		scanArgs := make([]interface{}, len(slaveCols))
-		for i := range scanArgs {
-			scanArgs[i] = &sql.RawBytes{}
-		}
-
-		if err := slaveStatusRows.Scan(scanArgs...); err != nil {
-			return err
-		}
-		for i, col := range slaveCols {
-			if value, ok := parseStatus(*scanArgs[i].(*sql.RawBytes)); ok { // Silently skip unparsable values.
-				ch <- prometheus.MustNewConstMetric(
-					newDesc(slaveStatus, strings.ToLower(col), "Generic metric from SHOW SLAVE STATUS."),
-					prometheus.UntypedValue,
-					value,
-				)
-			}
-		}
-	}
-	return nil
 }
 
 func scrapeInformationSchema(db *sql.DB, ch chan<- prometheus.Metric) error {
